refactor: CarRepository와 CarDetailsRepository를 분리
